pyproject.toml
[project]
name = "web-tieba-manager"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiosqlite>=0.21.0",
    "aiotieba>=4.6.0",
    "asyncmy>=0.2.10",
    "asyncpg>=0.30.0",
    "pydantic>=2.11.7",
    "pyyaml>=6.0.2",
    "sqlalchemy>=2.0.43",
]

[[tool.uv.index]]
url = "https://mirrors.aliyun.com/pypi/simple"
default = true
